# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DriveAssist API",
        "description": "API documentation for DriveAssist - Vehicle Diagnostics and Expert Marketplace",
        "contact": {
            "name": "DriveAssist Support",
            "email": "support@driveassist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/articles/categories/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles by category",
                "operationId": "2bf524fd7a19fc6cc349a3fb77da5175",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/articles/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article details",
                "operationId": "5116913f5b70bb0f7ab90575cfdbde20",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}